Moving search functionalities to SearchBox, fixing cors, testing connection with /stories endpoint
